relation teams and trips

+++ app/Http/Controllers/UsertripsController.php
@@ -32,7 +32,7 @@ class UsertripsController extends Controller
 
         parent::__construct();
         $this->model = new UserTrip();
-        $this->info  = $this->model->makeInfo('usertrips');
+        $this->info  = $this->model->makeInfo('Usertrip');
         $this->data  = array(
             'pageTitle'  => $this->info['title'],
             'pageNote'   => $this->info['note'],
@@ -350,6 +350,9 @@ class UsertripsController extends Controller
 
     public function compareRFP($rfp_id)
     {
+        if (!\Auth::check()) {
+            return redirect('user/login')->with('status', 'error')->with('message', 'You are no Logged in');
+        }
         $value = explode(',', $rfp_id);
 
         $rfps = Rfp::with('userInfo', 'userInfo.hotel')->whereIn("user_trip_id", $value)->get();
@@ -376,21 +379,18 @@ class UsertripsController extends Controller
          *
          * @var \App\Models\Rfp $rfp
          */
-        $rfp = Rfp::findOrFail($rfp_id);
 
-        if (!$user->can('accept', $rfp)) {
-            return response()->json([
-                'success'   => false,
-                'view_data' => 'Sorry, you are not allowed to accept this RFP.',
-            ]);
+        if (!\Auth::check()) {
+            return redirect('user/login')->with('status', 'error')->with('message', 'You are no Logged in');
         }
-
-        $rfp->update(['status' => Rfp::STATUS_BID_SELECTED]);
-
-        if ($rfp->user_id != 0) {
-            $receiver = $rfp->user;
-            $hotel    = $rfp->hotel;
-            $trip     = $rfp->trip;
+        Rfp::where('id', $rfp_id)->update(['status' => 2]);
+        $trip_id = Rfp::findOrFail($rfp_id);
+        
+        if ($trip_id->user_id != 0) {
+            $group = User::findOrFail($trip_id->user_id);
+            $hotel = Hotel::findOrFail($group->hotel_id);
+            $coordinatorId=Rfp::with('trip')->where('id',$rfp_id)->first();
+            //$log_id   = Session::get('uid');
 
             $agree_id = AgreementForm::where('for_rfp', $rfp_id)->first();
 
@@ -499,6 +499,7 @@ class UsertripsController extends Controller
     public function acceptAgree($rfp_id)
     {
 
+
         $trip            = Rfp::findOrFail($rfp_id);
         $trip_entry      = UserTrip::findOrFail($trip->user_trip_id);
         $trip_user       = User::findOrFail($trip->user_id);
@@ -574,6 +575,9 @@ class UsertripsController extends Controller
 
     public function declineRFP($rfp_id, $reason)
     {
+        if (!\Auth::check()) {
+            return redirect('user/login')->with('status', 'error')->with('message', 'You are no Logged in');
+        }
         Rfp::where('id', $rfp_id)->update(['status' => 3, 'decline_reason' => $reason]);
         $user_trip_id = Rfp::where('user_trip_id', $rfp_id)->pluck('user_trip_id');
         foreach ($user_trip_id as $item_new) {
@@ -622,6 +626,7 @@ class UsertripsController extends Controller
 
     public function getTeamview(Request $request)
     {
+
         $q                  = (new Team)->newQuery();
         $teams              = $q->orderBy('id', 'desc')->get();
         $user               = Auth::user();
@@ -637,6 +642,9 @@ class UsertripsController extends Controller
 
     public function getTeamdelete($id)
     {
+         if (!\Auth::check()) {
+            return redirect('user/login')->with('status', 'error')->with('message', 'You are no Logged in');
+        }
         Team::findOrFail($id)->delete();
         Session::flash("success", "Record Deleted");
 
@@ -645,6 +653,10 @@ class UsertripsController extends Controller
 
     public function uploadRoster($rfp_id, $team)
     {
+        if (!\Auth::check()) {
+            return redirect('user/login')->with('status', 'error')->with('message', 'You are no Logged in');
+        }
+
         Rfp::where('id', $rfp_id)->update(['team' => $team]);
 
         return response()->json([
@@ -655,13 +667,18 @@ class UsertripsController extends Controller
 
     public function uploadRosters($rfp_id)
     {
+         if (!\Auth::check()) {
+            return redirect('user/login')->with('status', 'error')->with('message', 'You are no Logged in');
+        }
         $teams = Rfp::where('id', $rfp_id)->get();
-
         return view('usertrips' . '.public' . '.uploadRoster', compact('teams'));
     }
 
     public function uploadRostertore(Request $request)
     {
+         if (!\Auth::check()) {
+            return redirect('user/login')->with('status', 'error')->with('message', 'You are no Logged in');
+        }
         $rfp_id          = $request->input('rfp_id');
         $file            = $request->file('room_file')->getClientOriginalName();
         $destinationPath = './uploads/users/';
@@ -679,22 +696,27 @@ class UsertripsController extends Controller
 
     public function show_trips()
     {
+
+         if (!\Auth::check()) {
+            return redirect(URL("/"));
+        }
+        //if (Session::get('level') != 4) {
         /** @var \App\User $user */
         $user = Auth::user();
 
         if (!$user->is_manager && !$user->is_subcoordinator) {
+
             return redirect(URL("/"));
         }
+        
 
-        if ($user->is_manager) {
+       /* if ($user->is_manager) {
             $trips = $user->organization->trips();
         } else {
             $trips = $user->trips();
-        }
+        }*/
 
-        $trips = $trips->with('tripuser', 'rfps')
-                       ->orderBy('added', 'desc')
-                       ->get();
+        $trips = Usertrip::with('tripuser', 'rfps')->where('entry_by',$user->id)->orderBy('added', 'desc')->get();
 
         $data_client = $user;
 
@@ -709,10 +731,12 @@ class UsertripsController extends Controller
         $get_invoice = Rfp::where("status", '!=', 3)->get();
         $data_accept = Rfp::where("status", 2)->get();
         $data_submit = Rfp::where("status", 1)->get();
-        $team        = Team::all();
 
+        //$team = Team::where('user_id',$user->id)->get();
+        $team        = Team::all();
         return view('coordinator.viewtrips',
             compact('trips', 'amenities', 'data_client', 'purchases', 'data_submit', 'get_invoice', 'data_accept', 'client', 'team'));
+
 
     }
